cpsw: add rx h/w ts to xdp_data buffer
On rx side it's Ok to add directly to the frame, but in case of tx,
when ts can be delayed, no way except error queue. By a big account
both of them could be in error queue or in descriptor field. Get ts only
for PTPl2 packets as no any reason on other types.

Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@linaro.org>